requirements.txt
Flask==1.0.2
pyOpenSSL==19.0.0
